requirements.txt
boto3>=1.34.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dateutil>=2.8.2 